+++ code/lib/hw-lib/HwKeypad.cpp
@@ -12,7 +12,7 @@ private:
     Button* confirmButton;
 
 public:
-    HwKeypad(int cancelPin, int downPin, int upPin, int confirmPin) {
+    HwKeypad(int8_t cancelPin, int8_t downPin, int8_t upPin, int8_t confirmPin) {
         cancelButton = new Button(cancelPin, INPUT_PULLUP, 250);
         downButton = new Button(downPin, INPUT_PULLUP, 250);
         upButton = new Button(upPin, INPUT_PULLUP, 250);
